back-end/conflu/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.6.0